Rename path for predefined special problems.
